requirements.txt
attrs==22.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.0
iniconfig==1.1.1
kiwisolver==1.4.9
markdown-it-py==3.0.0
matplotlib==3.10.7
mdurl==0.1.2
numpy==2.3.5
packaging==21.3
pandas==2.2.0
Pillow==9.3.0
playsound==1.3.0
pluggy==1.0.0
pyarrow==15.0.0
Pygments==2.17.2
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
rich==13.7.0
shellingham==1.5.4
six==1.16.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4